backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.111.0",
    "langchain-openai>=1.1.11",
    "langchain-pinecone>=0.2.13",
    "openai>=1.35.3",
    "pinecone-client>=4.1.1",
    "pydantic>=2.7.4",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.30.1",
]
